backend/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "batumi-lunch-backend"
version = "0.1.0"
description = "Batumi Lunch web API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "alembic>=1.13.1",
    "psycopg[binary]>=3.1.18",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.3",
    "rq>=1.15.1",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "boto3>=1.34.79",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt[crypto]>=2.8.0",
    "email-validator>=2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.5",
    "types-redis>=4.6.0.20240218"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --disable-warnings"
testpaths = [
    "backend/tests/unit",
    "backend/tests/integration",
]

[tool.coverage.run]
branch = true
source = ["backend/app"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

[tool.mypy]
python_version = "3.13"
packages = ["app"]
ignore_missing_imports = true
explicit_package_bases = true
